Refactor: move get_user in its own file. Rename to get_current_user. Return a response with the correct schema.
