+++ src/web/users/mod.rs
@@ -1,8 +1,10 @@
+mod current_user;
 mod login;
 mod register;
 mod responses;
 mod users;
 
+pub use current_user::get_current_user;
 pub use login::login;
 pub use register::register;
-pub use users::{get_user, update_user};
+pub use users::update_user;
